Logo Generator/Logo Generator/requirements.txt
replicate
requests
pillow
python-dotenv
google-generativeai
